Bug Fix for Unit Test
Unit Test was producing an error if there was a problem but not in the
way I desire. (R was producing it, not me). So I fixed that. Should
output correctly. Discovered when breaking the code that it doesn't say
exactly WHERE the code broke, so I'm going to fix that.